Create the remix contract
Problem Statement
When a vendor purchases vegetables in the APMC yard it goes through an IoT enabled pathway which bifurcates good and bad vegetables and gives the count. The output of the IoT (count of good and bad vegetables) will act as input to the DApp, which generates the following report:
1. Total vegetables purchased by the vendor
2. Net price the vendor has spent for all the vegetable
Assume that the vendor will inturn sells the vegetables to the customers with a cost of Rs. 20 + the net price of the vegetable. Show whether the vendor is in profit or loss.
Apart from these features, you can add your own features to make the portal user friendly.
Create a smart contract and suitable GUI for the same and use Ethereum test network to deploy and run the smart contract.